chore: Update npm dependencies
