+++ web/src/components/HomePageTwo/HomePageTwo.css
@@ -109,17 +109,19 @@
   padding: 1rem;
   text-decoration: none;
   color: black;
-  transition: all 0.3s ease;
-}
-
-.search_results a .selected {
-  background-color: #f1f1f1;
-  padding: 1rem;
-  padding-left: 0;
+  transition: all 0.2s ease;
 }
 
 .search_results a:hover {
   background-color: #f1f1f1;
+}
+
+.highlight {
+  background-color: red;
+}
+
+.search_results a .selected {
+  color: black;
 }
 
 .search_results:active {
@@ -192,13 +194,13 @@
 .InfoBarHome {
   display: flex;
   align-items: center;
-  justify-content: space-evenly;
   margin-top: 3rem;
   flex-wrap: wrap;
   gap: 2rem;
   overflow: hidden;
-  position: relative;
   height: 3rem; /* Set to height of a single item */
+  width: 100%;
+  max-width: 138rem;
 }
 
 .scrollContainer {
@@ -990,4 +992,26 @@
   100% {
     transform: translateY(0); /* Reset to initial position */
   }
+}
+
+.suggestions {
+  margin-top: 10px;
+  border: 1px solid #ccc;
+  background: #fff;
+  padding: 10px;
+  border-radius: 4px;
+}
+
+.suggestion_item {
+  padding: 5px;
+  cursor: pointer;
+}
+
+.suggestion_item:hover {
+  background: #f0f0f0;
+}
+
+.search_dropdown_item {
+  padding: 10px;
+  cursor: pointer;
 }
